oe-layertool-setup: fix dldir logic for templates
Previously templates were always being overridden with this internal
value that just so happened to be functionally equivalent to the
template DL_DIR.

Template:
	DL_DIR = "${OEBASE}/downloads"

Internal override:
	DL_DIR = "$oebase/downloads"

Functionally equivalent, but the internal override would be expanded to
include the full path of "$oebase" before being written to the file. In
the end it was unnecessarily bound to the context in which oe-layersetup
was executed.

This change lets us keep the local.conf sample/template value unless a
value is specified over cmdline or a backup value is found. It also
explicitly warns the user when a backup value is used.

Signed-off-by: Randolph Sapp <rs@ti.com>
Signed-off-by: Ryan Eatmon <reatmon@ti.com>