configs/arago-scarthgap-wip-config: Remove meta-arago-demos
Remove the meta-arago-demos layer to help break the dependency on
meta-qt5. This comes after a patch in meta-arago-distro that removed all
of the packages from meta-arago-demos. These were the last of the
"demos" in meta-arago and they were only being built for legacy devices.
Demos should primarily be in layers above arago unless they are used for
functionality testing.

These modifications are required as part of the effort to make arago
less dependent on meta-qt5. Moving forward TI builds of arago will
enable and test Qt6 by default, but unlike before this will not be done
in a way that creates an explicit dependency. This should allow
customers and layers that consume or extend arago to toggle Qt support
more easily.

The changes to scarthgap-next and scarthgap will come as the removal
patch in meta-arago makes it way through the CICD process.

Signed-off-by: Randolph Sapp <rs@ti.com>
Signed-off-by: Ryan Eatmon <reatmon@ti.com>