templates: Fix paths for DL_DIR and SSTATE_DIR
When the conversion to the new XML/XSLT system was done, the reference
to OEBASE was converted to TOPDIR.  The value for OEBASE was
historically the directory above the bitbake run dir which is what
TOPDIR is set to.  This made the downloads and sstate_cache directories
appear down a level of directory that was unintended.

This change updates the values for DL_DIR and SSTATE_DIR to be relative
to TOPDIR but back in the same location they were originally meant to be
in.

Signed-off-by: Ryan Eatmon <reatmon@ti.com>